+++ CS/Solution/Default.aspx
@@ -1,11 +1,7 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="Default.aspx.cs" Inherits="_Default" %>
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="Default.aspx.cs" Inherits="Solution.Default" %>
 
-<%@ Register Assembly="DevExpress.Web.v21.2, Version=21.2.5.0, Culture=neutral, PublicKeyToken=b88d1754d700e49a"
-    Namespace="DevExpress.Web" TagPrefix="dx" %>
+<!DOCTYPE html>
 
-
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.1//EN" "http://www.w3.org/TR/xhtml11/DTD/xhtml11.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml">
 <head runat="server">
     <title>How to show ASPxLoadingPanel during export</title>
@@ -33,6 +29,7 @@
             <dx:ASPxGridView ID="MyGrid" runat="server" EnableCallbackAnimation="true" AutoGenerateColumns="true">
                 <SettingsExport EnableClientSideExportAPI="true"/>
             </dx:ASPxGridView>
+
             <div style="display: flex; flex-direction: row; justify-content: flex-start">
                 <dx:ASPxButton ID="btnExportPDF" runat="server" AutoPostBack="false" Text="Export to PDF">
                     <ClientSideEvents Click="function(s,e){onExportWithCallbackClick('pdf')}" />
@@ -45,7 +42,6 @@
                 </dx:ASPxButton>
             </div>
             
-
             <dx:ASPxButton ID="response_btn" runat="server" ClientInstanceName="response_btn" ClientVisible="false"
                 OnClick="response_btn_Click" />
             <dx:ASPxLoadingPanel ID="ASPxLoadingPanel1" runat="server" ClientInstanceName="loadingPanel"></dx:ASPxLoadingPanel>
