+++ CS/Solution/Global.asax
@@ -1,0 +1,1 @@
+<%@ Application Codebehind="Global.asax.cs" Inherits="Solution.Global_asax" Language="C#" %>
